fixed TestPut
